qa-testdefinitions: Split can tests to avoid failures
Split the can tests in two, one for basic tests that doesn't need the
two can interfaces to be related, and the other one is for a board with
two related interfaces.

Modify the test_can_modules to be run on a board with only one can if at
least.

AGL-bug: SPEC-2878

Change-Id: I6f761458c3e191aa8000d19cb05b0592566cb6c9
Signed-off-by: Khouloud Touil <ktouil@baylibre.com>