SPEC-3427: check if RID is null
RID could be null if process failed to start.
In that case, we will retry to start it.
On a second fail, we will not continue the test suite and fail.

In th same time add a second status check, but after a delay to be sure
the service is still working.

Bug-AGL: SPEC-3427
Change-Id: I566f1aefa06cbb791674373340ce87be220b18a2
Signed-off-by: Corentin LABBE <clabbe@baylibre.com>