agl-screenshot-test.sh: Redo the compositor restart
Stopping services and the client shell before restarting the compositor,
would avoid some large delays until the new compositor instance
is brought up.

With this approach, were we stop explicitly the client shell and the
launcher, we're basically have a hard stop, and should avoid those huge
delays for qemu-arm. Until we can confirm this is the case, we still
maintain that delay.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ifa067ac40de162291d7d6cfb70035e87add07220